⚡ Prevent unnecessary image loadings on tab change
Changing tab caused a recompose of the PhotoPage, which
caused the re-creation of the ImageRequest.
Thanksfully, Coil was caching it, but still it caused
couple hundreds of ms where Loading was displayed.

With this state hoisting, the state is kept at RoundPlayPage
and will be kept throughout the tab changes